backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
httpx==0.27.0
pydantic==2.7.0
python-dotenv==1.0.1
sse-starlette==2.1.0
tenacity==8.3.0
tavily-python==0.3.3